pyproject.toml
[tool.poetry]
name = "itss"
version = "0.1.0"
description = ""
authors = ["tranduchoangnam <20febnam@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.108.0"}
uvicorn = {extras = ["standard"], version = "^0.25.0"}
pydantic = "^2.5.3"
sqlalchemy = "^2.0.24"
alembic = "^1.13.1"
sqlmodel = "^0.0.14"
fastapi-mail = {extras = ["httpx"], version = "^1.4.1"}
boto3 = "^1.34.91"
pytz = "^2024.1"
pandas = "^2.2.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
psycopg2 = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
